+++ src/Contexts/Adaptors/EVMAdaptors/EVMHomeAdaptorProvider.tsx
@@ -1,7 +1,7 @@
 import React from "react";
 import { Bridge, BridgeFactory } from "@chainsafe/chainbridge-contracts";
-import { useWeb3 } from "@chainsafe/web3-context";
-import { BigNumber, utils } from "ethers";
+import { useConnectWallet, useSetChain } from "@web3-onboard/react";
+import { BigNumber, utils, ethers} from "ethers";
 import { useCallback, useEffect, useState } from "react";
 import {
   chainbridgeConfig,
@@ -20,25 +20,94 @@ import { decodeAddress } from "@polkadot/util-crypto";
 import { getPriceCompatibility } from "./helpers";
 import { hasTokenSupplies } from "../SubstrateApis/ChainBridgeAPI";
 import { ApiPromise } from "@polkadot/api";
-import { localStorageVars } from "../../../Constants/constants";
-const { ONBOARD_SELECTED_WALLET } = localStorageVars;
 
 export const EVMHomeAdaptorProvider = ({
   children,
 }: IHomeBridgeProviderProps) => {
-  const {
-    isReady,
-    network,
-    provider,
-    gasPrice,
-    address,
-    tokens,
-    wallet,
-    checkIsReady,
-    ethBalance,
-    onboard,
-    resetOnboard,
-  } = useWeb3();
+  const [
+    {
+      wallet, // the wallet that has been connected or null if not yet connected
+      connecting // boolean indicating if connection is in progress,
+    },
+    connect, // function to call to initiate user to connect wallet
+    disconnect, // function to call with wallet<DisconnectOptions> to disconnect wallet
+    updateBalances, // function to be called with an optional array of wallet addresses connected through Onboard to update balance or empty/no params to update all connected wallets
+    setWalletModules, // function to be called with an array of wallet modules to conditionally allow connection of wallet types i.e. setWalletModules([ledger, trezor, injected])
+    setPrimaryWallet // function that can set the primary wallet and/or primary account within that wallet. The wallet that is set needs to be passed in for the first parameter and if you would like to set the primary account, the address of that account also needs to be passed in
+  ] = useConnectWallet();
+
+  const [
+    {
+      chains, // the list of chains that web3-onboard was initialized with
+      connectedChain, // the current chain the user's wallet is connected to
+      settingChain // boolean indicating if the chain is in the process of being set
+    },
+    setChain // function to call to initiate user to switch chains in their wallet
+  ] = useSetChain();
+
+  // Tokens are records of the address and token information for the currently selected chain.
+  const [tokens, setTokens] = useState<Record<TokenConfig["address"], TokenConfig>>({});
+
+  const [ethBalance, setEthBalance] = useState(0);
+
+  const [provider, setProvider] = useState<ethers.providers.Web3Provider>();
+
+  // State for wallet readiness status. Initially false.
+  const [isReady, setIsReady] = useState(false);
+
+  // State for the connected wallet address. Initially undefined.
+  const [address, updateAddress] = useState<string>();
+
+  // Network is an id of current chain.
+  const [network, setNetwork] = useState<number>();
+
+  const fetchGasPrice = useCallback(async () => {
+    if (!provider) return null;
+    try {
+      const fetchedGasPrice = await provider.getGasPrice();
+      return Number(fetchedGasPrice);
+    } catch (err) {
+      console.error(err);
+    }
+  }, [provider]);
+
+  useEffect(() => {
+    if (wallet?.provider) {
+      // if using ethers v6 this is:
+      // ethersProvider = new ethers.BrowserProvider(wallet.provider, 'any')
+      setProvider(new ethers.providers.Web3Provider(wallet.provider, 'any'))
+    } else {
+      setProvider(undefined);
+    }
+  }, [wallet]);
+
+  // Setup eth balance.
+  useEffect(() => {
+    // Active account is always the first in wallet.
+    const ethBalanceInWallet = wallet?.accounts?.[0]?.balance?.ETH;
+
+    if (ethBalanceInWallet) {
+      setEthBalance(Number(ethBalanceInWallet));
+    }
+  }, [wallet]);
+
+  useEffect(() => {
+    setIsReady(Boolean(wallet) && !connecting);
+  }, [wallet, connecting]);
+
+  useEffect(() => {
+    updateAddress(wallet?.accounts[0].address);
+  }, [wallet]);
+
+  useEffect(() => {
+    setNetwork(Number(connectedChain?.id));
+  }, [connectedChain]);
+
+  const disconnectWallet = () => {
+    if (wallet) {
+      disconnect(wallet);
+    }
+  };
 
   const {
     homeChainConfig,
@@ -81,10 +150,8 @@ export const EVMHomeAdaptorProvider = ({
   const [account, setAccount] = useState<string | undefined>();
 
   const checkWallet = useCallback(async () => {
-    let success = false;
     try {
-     success = await checkIsReady();
-      if (success) {
+      if (wallet) {
         if (homeChainConfig && network && isReady && provider) {
           const signer = provider.getSigner();
           if (!signer) {
@@ -120,9 +187,9 @@ export const EVMHomeAdaptorProvider = ({
       console.error(err);
     } finally {
       setInitialising(false);
-      return success;
-    }             
-  }, [checkIsReady, homeChainConfig, isReady, network, provider]);
+      return Boolean(wallet);
+    }
+  }, [wallet, homeChainConfig, isReady, network, provider]);
 
   useEffect(() => {
     if (network) {
@@ -131,35 +198,43 @@ export const EVMHomeAdaptorProvider = ({
       const supported = !!chainbridgeConfig.chains.find(chain => chain.networkId === network);
       setNetworkSupported(supported);
       if (!!network && !supported) {
-        resetOnboard();
+        disconnectWallet();
         setWalletType("unset");
       }
       if (chain) {
+        // `setTokens` maps over the chain tokens. For each token, it checks if it exists in the wallet.
+        // If it does, it uses the wallet's balance. If it doesn't, it sets balance as "0".
+        // The function constructs an object where each key is a token address and its value is the token details and balance.
+        setTokens(chain.tokens.reduce((acc, currentToken) => {
+          const matchingTokenInWallet = wallet?.accounts[0]?.secondaryTokens?.find(
+              (walletToken) => walletToken.name === currentToken.symbol
+          );
+
+          // Calculate balance by converting from 'wei' to 'ether' and then to the token's units
+          // If `matchingTokenInWallet` does not exist, set balance to "0"
+          const balance = matchingTokenInWallet
+              ? +matchingTokenInWallet.balance * Math.pow(10, 18) / Math.pow(10, currentToken.decimals || 10)
+              : "0";
+
+          return {
+            ...acc,
+            [currentToken.address]: {
+              ...currentToken,
+              balance,
+            },
+          };
+        }, {}));
         handleSetHomeChain(chain.chainId);
+      } else {
+        setTokens({});
       }
     }
-  }, [handleSetHomeChain, homeChains, network, setNetworkId]);
+  }, [handleSetHomeChain, homeChains, network, setNetworkId, wallet?.accounts]);
 
   useEffect(() => {
-    if (initialising || homeBridge || !onboard) return;
+    if (initialising || homeBridge) return;
     console.log("Starting init");
     setInitialising(true);
-
-    wallet?.provider?.on("error", (err: any) => {
-      console.error("Wallet provider error:", err);
-    });
-    
-    // On the first connect to a blockchain this event doesn't happen
-    wallet?.provider?.on("chainChanged", (newNetworkId: number) => {
-      console.log("Chain changed:", { networkId, newNetworkId });
-      if (newNetworkId === networkId) return;
-      setNetworkId(
-        newNetworkId.toString().substring(0, 2) === '0x' 
-        ? parseInt(newNetworkId.toString(), 16)
-        : newNetworkId 
-      );
-      if (isReady && networkSupported) window.location.reload();
-    });
 
     wallet?.provider?.on("accountsChanged", (accounts: string[])=> {
       console.log("Accounts changed:", { walletSelected, account, accounts });
@@ -168,46 +243,21 @@ export const EVMHomeAdaptorProvider = ({
       setAccount(accounts[0].toLowerCase())
     });
 
-    const selectedWallet = localStorage.getItem(ONBOARD_SELECTED_WALLET) as string;
-    let connected = false;
-
     if (walletType === "Ethereum") {
-      onboard
-          .walletSelect(selectedWallet)
-          .then(async (success) => {
-            console.log("Wallet select:", { success });
-            setWalletSelected(success);
-            if (success) {
-              connected = await checkWallet() as boolean;
-              console.log("Wallet check:", { connected });
-            }
-          })
-          .catch((error) => {
-            console.error(error);
-            setInitialising(false);
-            connected = false;
-          })
-          .finally(() => {
-            if (!connected) {
-              resetOnboard();
-              setWalletType("unset");
-            }
-            console.log("Wallet connection finished:", { connected });
-          })
-      }
+      connect().then(() => {
+        checkWallet();
+      });
+    }
   }, [
     checkWallet,
     initialising,
     homeChainConfig,
     isReady,
     provider,
-    checkIsReady,
     network,
     networkId,
     setNetworkId,
     homeBridge,
-    onboard,
-    resetOnboard,
     walletSelected,
     setWalletType,
     account,
@@ -238,11 +288,8 @@ export const EVMHomeAdaptorProvider = ({
   }, [homeBridge]);
 
   const handleConnect = useCallback(async () => {
-    if (wallet && wallet.connect && network) {
-      await onboard?.walletSelect();
-      await wallet.connect();
-    }
-  }, [wallet, network, onboard]);
+   await connect();
+  }, [connect]);
 
   const handleCheckSupplies = useCallback(
     async (amount: number) => {
@@ -327,6 +374,12 @@ export const EVMHomeAdaptorProvider = ({
         recipient.substr(2); // recipientAddress (?? bytes)
 
       try {
+        const gasPrice = await fetchGasPrice();
+
+        if (!gasPrice) {
+          throw new Error('Failed to fetch gas price');
+        }
+
         const gasPriceCompatibility = await getPriceCompatibility(
           provider,
           homeChainConfig,
@@ -428,7 +481,6 @@ export const EVMHomeAdaptorProvider = ({
       address,
       bridgeFee,
       homeChainConfig,
-      gasPrice,
       provider,
       setDepositNonce,
       setTransactionStatus,
@@ -441,6 +493,12 @@ export const EVMHomeAdaptorProvider = ({
       return "not ready";
 
     try {
+      const gasPrice = await fetchGasPrice();
+
+      if (!gasPrice) {
+        throw new Error('Failed to fetch gas price');
+      }
+
       const gasPriceCompatibility = await getPriceCompatibility(
         provider,
         homeChainConfig,
@@ -469,6 +527,12 @@ export const EVMHomeAdaptorProvider = ({
       return "not ready";
 
     try {
+      const gasPrice = await fetchGasPrice();
+
+      if (!gasPrice) {
+        throw new Error('Failed to fetch gas price');
+      }
+
       const gasPriceCompatibility = await getPriceCompatibility(
         provider,
         homeChainConfig,
@@ -496,7 +560,7 @@ export const EVMHomeAdaptorProvider = ({
       value={{
         connect: handleConnect,
         disconnect: async () => {
-          resetOnboard();
+          disconnectWallet()
           setWalletType("unset");
         },
         bridgeFee,