+++ CHANGELOG.md
@@ -1,7 +1,7 @@
 # CHANGELOG
 
 ## vNext
-...
+- Migrated to WalletConnect v2
 
 ## v0.23.0
 - Disabled WalletConnect v1