+++ CHANGELOG.md
@@ -3,6 +3,9 @@
 ## vNext
 ...
 
+## v0.24.0
+- Migrated to WalletConnect v2
+
 ## v0.23.0
 - Disabled WalletConnect v1